% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_patents.R
\name{download_patents}
\alias{download_patents}
\title{Retrieve Full Patent Text}
\usage{
download_patents(ids = NULL, outFile = NULL, ids_type = "application",
  query = NULL, inventionTitle = NULL,
  inventionSubjectMatterCategory = NULL, assigneeEntityName = NULL,
  fromDate = NULL, toDate = NULL, filingDateFromDate = NULL,
  filingDateToDate = NULL, inventorNameText = NULL, claimText = NULL,
  abstractText = NULL, descriptionText = NULL, sortField = NULL,
  ascending = FALSE, grant = FALSE, limit = FALSE, series = "US",
  verbose = FALSE)
}
\arguments{
\item{ids}{A vector of application numbers, patent numbers, or document IDs (e.g., \code{"US9748552"}).}

\item{outFile}{Path to a \code{.csv} file to write results to.}

\item{ids_type}{Specifies what \code{ids} are, between \code{applications} (default), \code{patents},
\code{publication} (for publication document IDs), or \code{grant} (for grant document IDs).}

\item{query}{Search query; either a term/phrase (\code{searchText}) or set of Boolean criteria (\code{criteriaSearchText}).}

\item{inventionTitle}{Invention title, as it appears on the first page of the specification.}

\item{inventionSubjectMatterCategory}{Patent category (e.g., \code{DESIGN}).}

\item{assigneeEntityName}{Name of the patent assignee.}

\item{fromDate, toDate, filingDateFromDate, filingDateToDate}{Publication/grant and/or filing date ranges
in \code{YYYY-MM-DD} format (e.g., \code{2019-01-20}).}

\item{inventorNameText, claimText, abstractText, descriptionText}{Text to search for in specific portions;
\code{query} will search in all portions if it is a term or phrase.}

\item{sortField}{Field by which to sort results; defaults to date.}

\item{ascending}{Logical; if \code{TRUE}, will sort in ascending order.}

\item{grant}{Logical; if \code{TRUE}, searches for grants rather than publications.}

\item{limit}{Maximum number of results to return.}

\item{series}{Series code to prepend to \code{ids} if they are application or patent numbers and not already included.}

\item{verbose}{Logical; if \code{TRUE}, will print status messages.}
}
\value{
Path to the downloaded package, if it exists and \code{await} was a number; otherwise \code{NULL}.
}
\description{
Search for and return a set of full patent texts.
}
\examples{
\dontrun{
# retrieve a specific set of patents:
patents <- download_patents(c(16978600, 17087826, 16041592, 16727940, 17070594))
}
}
