% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_patentsview_bulk.R
\name{download_patentsview_bulk}
\alias{download_patentsview_bulk}
\title{Download PatentsView Bulk}
\usage{
download_patentsview_bulk(table, dir = tempdir(), pregrant = FALSE,
  partition = NULL, make_db = FALSE, format = "parquet",
  return_table = TRUE, overwrite = FALSE)
}
\arguments{
\item{table}{Name of the table to download, as listed on the
\href{https://patentsview.org/download/data-download-tables}{Data Downloads page}.}

\item{dir}{Directory in which to save original tables and the dataset if \code{make_db} is \code{TRUE}.}

\item{pregrant}{Logical; if \code{TRUE}, will download the pre-grant version of \code{table}.}

\item{partition}{A vector of column names to be used as partition keys, if \code{make_db} is \code{TRUE}.
Can be a named list of functions, which will be used to create new columns, then partition by those columns
(e.g., \code{list(series_code = function(d) substr(d$patent_id, 1, 2))}). If \code{NULL} not specified, will choose
a column with the number of unique values closest to 20.}

\item{make_db}{Logical; if \code{TRUE}, will make an Arrow dataset out of the downloaded table.}

\item{format}{Format of the dataset, if \code{make_db} is \code{TRUE}.}

\item{return_table}{Logical; if \code{FALSE}, returns the path to the file, rather than the read-in table.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing files (raw and prepared).}
}
\value{
The original table (if \code{return_table} is \code{TRUE}; as a \code{tibble}), an opened dataset
(if the path to the dataset exists), or the path to the downloaded file.
}
\description{
Download one of the bulk tables from \href{https://patentsview.org}{PatentsView}, and optionally
make an Arrow dataset from it.
}
\examples{
\dontrun{
# download the application table of granted patents
download_patentsview_bulk(".", "application", partition = "series_code")
}
}
