% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_peds.R
\name{download_peds}
\alias{download_peds}
\title{Download Patent Examination Data}
\usage{
download_peds(searchText = "*:*", filters = list("*:*"), outFile = NULL,
  start = 0, minMatch = "100\%", overwrite = FALSE, waits = 50,
  wait = 10, endpoint = "https://ped.uspto.gov/api/queries/",
  compress = TRUE, load = TRUE, verbose = TRUE)
}
\arguments{
\item{searchText}{Text to search for; \code{"*:*"} for none (default); see
the FAQ on the \href{https://ped.uspto.gov/peds}{PEDS site}.}

\item{filters}{A list of characters representing filter conditions in the form of \code{"field:value"}
(e.g., \code{list('appStatus:"Patented Case"')}).}

\item{outFile}{Name of the final JSON file, or directory in which to save files with parameter-hash names.}

\item{start}{Initial record.}

\item{minMatch}{Minimum number of terms required to match.}

\item{overwrite}{Logical; overwrite any previous queries with the same body hash.}

\item{waits}{Number of times to check for a completed bundle before giving up.}

\item{wait}{Number of seconds to wait between retries (how long \code{waits} are).}

\item{endpoint}{PEDs API endpoint.}

\item{compress}{Logical; if \code{FALSE}, will not xz-compress the \code{outFile}.}

\item{load}{Logical; if \code{FALSE}, will not return the downloaded content, or load in existing content.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with an entry for each year in the set of patent metadata.
}
\description{
Download a bundled query response from the \href{https://ped.uspto.gov/peds}{Patent Examination Data System}.
}
\examples{
\dontrun{
# like case 3 in the API Documentation > API Tutorial on https://ped.uspto.gov/peds
download_peds(
  "firstNamedApplicant:(Google)",
  list('appStatus:"Patented Case"', "appFilingDate:[2013-01-01T00:00:00Z TO 2013-03-31T23:59:59Z]")
)
}
}
