% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_event_timing.R
\name{extract_event_timing}
\alias{extract_event_timing}
\title{Extract Timings from Examination Events}
\usage{
extract_event_timing(record, spans = list(`Initial Classification` =
  list(start = "^PGPC$", end = "^(?:OI|PG$)", before = "^DOCK$"),
  `Art Unit to Examiner` = list(start = "^DOCK$", end = "^FWDX$", before =
  "^(?:CTFR|EX\\\\.R|N/=\\\\.)"), `Examiner to Action` = list(start = "^FWDX$",
  end = "^(?:CT|EX\\\\.R|N/=\\\\.)", before = "^(?:DOCK$|A)")))
}
\arguments{
\item{record}{A matrix-like object with dates in the first column (recognizable by
\code{\link{as.Date}}) and event codes (recognizable by the patterns in \code{span}) in the second.}

\item{spans}{A list with entries defining each event window. Each entry should be a list with
entries for \code{start}, \code{end}, and \code{before}, each containing regular expression
strings identifying event codes.}
}
\value{
A list with a \code{data.frame} for each entry in \code{span}.
}
\description{
Extract timings between select event codes within an examination timeline.
}
\examples{
extract_event_timing(data.frame(
  date = c("2020-01-15", "2020-01-20"),
  event = c("FWDX", "CTNF")
))
}
