% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_examiner_name.R
\name{extract_examiner_name}
\alias{extract_examiner_name}
\title{Get Examiner Name from Office Actions}
\usage{
extract_examiner_name(office_action, collapse = TRUE)
}
\arguments{
\item{office_action}{A list of office action, as retrieved with \code{\link{download_office_actions}};
each entry should be a list with entries for \code{bodyText} (the office action text) and
\code{examinerEmplyeeNumber}.}

\item{collapse}{Logical; if \code{FALSE}, will return a row for every entered office action.}
}
\value{
A character vector with the Examiner's employee number and two possible names: one
extracted from the language (\code{intext}) and another from the sign-off (\code{signed}). The name
will be \code{NA} if there is no examiner (such as if a non-examiner submitted the action), or
if the format is not as expected. If \code{office_action} is a list of multiple office actions,
these will make up rows of a \code{data.frame}, which will be as many as the length of
\code{office_actions} if \code{collapse} is \code{FALSE}, or the length of unique IDs.
}
\description{
Extract the examiner's name from office action body text.
}
\examples{
\dontrun{
# retrieve the office actions associated with a particular application
office_actions <- download_office_actions("patentApplicationNumber:13877637")

# get the name of the examiner associated with the first action:
extract_examiner_name(office_actions[[1]])
}
}
