% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inventors.R
\name{extract_inventors}
\alias{extract_inventors}
\title{Get inventor information}
\usage{
extract_inventors(input)
}
\arguments{
\item{input}{A \code{data.frame} of documents (as returned from \code{\link{uspto_download}})
or a list of examination records (as returned from \code{\link{download_peds}}).}
}
\value{
A \code{data.frame} with inventor information.
}
\description{
Extract inventor information from patent/application documents or examination records.
}
\examples{
\dontrun{
# download an application
application <- uspto_download("US-20040216465-A1")

# get inventor information from it
extract_inventors(application)

# get the same from examination record
extract_inventors(download_peds("applId:10612573"))
}
}
