% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_info.R
\name{get_class_info}
\alias{get_class_info}
\title{Get Patent Classification Information}
\usage{
get_class_info(code, dir = tempdir())
}
\arguments{
\item{code}{A vector of USPC classification symbols (e.g., \code{D14}).}
}
\value{
A list with an entry for \code{class} and \code{description} for the entered code, and \code{subclasses}
containing a \code{data.frame} with a \code{subclass} (the subclass code) and \code{description} column.
}
\description{
Retrieve U.S. patent Class definitions and subclasses from
\href{https://www.uspto.gov/web/patents/classification}{USPTO Classification Resources}.
}
\examples{
\dontrun{
classifications <- get_class_info(c(428, 429))
}
}
