% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patentsview_category_matrix.R
\name{patentsview_class_matrix}
\alias{patentsview_class_matrix}
\title{Make Document-Class Matrix}
\usage{
patentsview_class_matrix(table, outFile = NULL, class_id = NULL,
  doc_id = "patent_id", rank = "sequence", type = NULL, ...,
  sparse = TRUE, overwrite = FALSE)
}
\arguments{
\item{table}{A matrix-like object with columns specified by \code{class_id}, \code{doc_id}, and \code{rank}.
Alternatively, the path to a tab-separated file containing such a matrix, an \code{Arrow} dataset, or the name
of a table, to be passed to \code{\link{download_patentsview_bulk}}.}

\item{outFile}{Path to an rds file to save the results to.}

\item{class_id, doc_id, rank}{Column names of the class, document ID, and class rank to be pulled from \code{table}.}

\item{type}{Table name, used to set a default \code{class_id}.}

\item{...}{Additional arguments to be passed to \code{\link{download_patentsview_bulk}},
if \code{table} is a table name.}

\item{sparse}{Logical; if \code{FALSE}, returns a regular, dense matrix.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites an existing \code{outFile} rather than loading it.}
}
\value{
A sparse matrix (or regular matrix if \code{sparse} is \code{FALSE}) with documents
in rows, classes in columns, and the class rank (sequence) as values.
}
\description{
Create A document-class matrix from a patentsview bulk table.
}
\examples{
table <- data.frame(
  patent_id = c("a", "a", "b"),
  class = c(1, 3, 2),
  sequence = c(1, 0, 0)
)
patentsview_class_matrix(table, class_id = "class")

\dontrun{

# get a matrix of WIPO class assignments
wipo_fields <- patentsview_class_matrix("wipo")

# get a subset without creating the full matrix
wipo <- download_patentsview_bulk("wipo", make_db = TRUE)
wipo_fields_sub <- patentsview_class_matrix(dplyr::compute(dplyr::filter(
  wipo, patent_id \%in\% c("10000002", "10000015", "10000017")
)))
}
}
