% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fulltext.R
\name{search_fulltext}
\alias{search_fulltext}
\title{Search for Patents or Patent Applications (legacy)}
\usage{
search_fulltext(query, applications = TRUE, limit = NULL, retries = 20,
  outdir = NULL, overwrite = FALSE, cores = parallel::detectCores() - 2,
  verbose = TRUE)
}
\arguments{
\item{query}{A query string. This can be a set of terms (e.g., \code{"term1 OR term2"}) or
an exact phrase (e.g., \code{'"a phrase"'}). To search within a particular
\href{https://appft.uspto.gov/netahtml/PTO/help/helpflds.html}{field}, prefix with the field code
and a forward slash (e.g., \code{"CPCL/G01W"}, which searches for "G01W" in the CPCL field).}

\item{applications}{Logical; specifies whether to search in the application database
(AppFT; \code{TRUE}; default) or the patent database (PatFT; \code{FALSE}).}

\item{limit}{A limit to the number of results retrieved, in steps of 50. By default, will
try to retrieve all results.}

\item{retries}{Number of retries for a given results page.}

\item{outdir}{Path to a directory in which to save results. Only writes if specified.}

\item{overwrite}{Logical; if \code{TRUE}, will overwrite caches results.}

\item{cores}{Number of CPU cores to use when reading in multiple pages.}

\item{verbose}{Logical; if \code{FALSE}, does not print status messages.}
}
\value{
A \code{data.frame} with columns for \code{index} (index of the result),
\code{id} (patent number or application document number), and \code{title} (title of patent or application).
}
\description{
Searches the Application (AppFT) or Patent (PatFT)
\href{https://appft.uspto.gov}{Full-Text Databases}, and returns
document/patent numbers and titles.
}
\examples{
\dontrun{
# search for applications with a meteorology classification:
# https://www.uspto.gov/web/patents/classification/cpc/html/cpc-G01W.html
results <- search_fulltext("CPCL/G01W")

# search for patents by date and title keyword
results <- search_fulltext("isd/1/8/2002 and motorcycle", FALSE)
}
}
\seealso{
For the newer search mechanism, see \code{\link{uspto_search}}.
}
