% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspto-package.R
\docType{package}
\name{uspto-package}
\alias{uspto}
\alias{uspto-package}
\title{uspto: U.S. Patent and Trademark Office API}
\description{
Offers functions to interact with U.S. Patent and Trademark Office public data.
}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Authors:
\itemize{
  \item Biocomplexity Institute [copyright holder]
}

}
\keyword{internal}
