% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspto_download.R
\name{uspto_download}
\alias{uspto_download}
\title{Download Patents or Patent Applications}
\usage{
uspto_download(guid, outDir = tempdir(), type = "US-PGPUB",
  cores = detectCores() - 2, compress = TRUE, verbose = FALSE)
}
\arguments{
\item{guid}{Vector of document numbers (e.g., \code{US-20220230294-A1}). Can also be a \code{data.frame} of results,
as returned from \code{\link{uspto_search}}; must have a \code{guid} column containing document numbers,
and a \code{type} column with the source database abbreviation (e.g., \code{US-PGPUB}).}

\item{outDir}{Path to a directory in which to save each individual document. If not specified, these are
saved to a temporary directory.}

\item{type}{A vector the same length as \code{guid} (repeated as necessary), indicating the source database of
the \code{guid}. Defaults to \code{US-PGPUB} for every ID.}

\item{cores}{Number of CPU cores to split requests across.}

\item{compress}{Logical; if \code{FALSE}, will not xz-compress each file.}

\item{verbose}{Logical; if \code{FALSE}, does not print status messages.}
}
\value{
A \code{data.frame} with a row for each document.
}
\description{
Downloads from the U.S. Pre-Grant Publications (US-PGPUB), Patents (USPAT), and/or
Optical Character Recognition (USOCR) databases. See the \href{https://ppubs.uspto.gov/pubwebapp/static/pages/landing.html}{web app}
to search for and view documents from the same source.
}
\examples{
\dontrun{
# start with a search
results <- uspto_search("G01W.CPCL.", "US-PGPUB")

# then download those results
applications <- uspto_download(results)
}
}
\seealso{
You can more efficiently download granted patents using \code{\link{download_patents}}.
}
