% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspto_search.R
\name{uspto_search}
\alias{uspto_search}
\title{Search for Patents or Patent Applications}
\usage{
uspto_search(query, databases = c("US-PGPUB", "USPAT", "USOCR"),
  outFile = NULL, limit = FALSE, sort = "date_publ desc",
  english_only = TRUE, spellCheck = FALSE, plurals = TRUE,
  britishEquivalents = TRUE, verbose = FALSE)
}
\arguments{
\item{query}{A query string; for example \code{"noodle"}, \code{"photographic AND noodle"},
\code{"wet adj4 silicon"}, or \code{"G01W.CPCL."}. See the Patent Public Search
\href{https://ppubs.uspto.gov/pubwebapp/static/pages/searchable-indexes.html}{training materials}.}

\item{databases}{A character vector specifying which databases to search in; a selection from
\code{c("US-PGPUB", "USPAT", "USOCR")}.}

\item{outFile}{Name of a \code{.csv} file to save results to. If this file exists, it will
be loaded instead of searching.}

\item{limit}{Maximum number of results to return; defaults to all results.}

\item{sort}{How to sort results; defaults to publication date (\code{"date_publ desc"}).}

\item{english_only}{Logical; if \code{FALSE}, will return patent and/or applications of any language.}

\item{spellCheck}{Logical; if \code{TRUE}, will spellcheck \code{query} terms.}

\item{plurals}{Logical; if \code{TRUE}, will look for plural and singular forms of \code{query} terms.}

\item{britishEquivalents}{Logical; if \code{TRUE}, will look for British and American forms of \code{query} terms.}

\item{verbose}{Logical; if \code{FALSE}, does not print status messages.}
}
\value{
A \code{data.frame} of results, if any were found; otherwise \code{NULL}.
}
\description{
Searches the U.S. Pre-Grant Publications (US-PGPUB), Patents (USPAT), and/or
Optical Character Recognition (USOCR) databases. See the \href{https://ppubs.uspto.gov/pubwebapp/static/pages/landing.html}{web app}
for an interactive search.
}
\examples{
\dontrun{
# search for applications with a meteorology classification:
# https://www.uspto.gov/web/patents/classification/cpc/html/cpc-G01W.html
results <- uspto_search("G01W.CPCL.", "US-PGPUB")

# search by date and keyword:
results <- uspto_search("20020801.pd. AND motorcycle")
}
}
