% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspto_search_patentsview.R
\name{uspto_search_patentsview}
\alias{uspto_search_patentsview}
\title{Search PatentsView}
\usage{
uspto_search_patentsview(query, fields = NULL, sort = NULL,
  outDir = tempdir(), include_related = FALSE, start = 1,
  limit = 10000, endpoint = "patent", post = TRUE, retry = 10,
  verbose = FALSE, legacy = TRUE, key = Sys.getenv("PATENTSVIEW_KEY"),
  cores = detectCores() - 1, overwrite = FALSE)
}
\arguments{
\item{query}{A query string or list; for example \code{list(patent_number = 7861317)} or
\code{'{"_text_all": {"patent_abstract": ["photographic", "noodle"]}}'}. See
\href{https://patentsview.org/apis/api-query-language}{API Query Language}.}

\item{fields}{A vector of fields to return; defaults to all for the given endpoint. Set to \code{NULL}
to use the API's default.}

\item{sort}{A list or list of lists specifying how to sort results (e.g., \code{list(patent_date = "asc")}).}

\item{outDir}{Directory in which to save results; defaults to a temporary directory.}

\item{include_related}{Logical; if \code{TRUE}, includes subentities related to entities matching the search criteria.}

\item{start}{Page to start collecting results on.}

\item{limit}{Maximum number of results to return; sets page size, up to 10,000 per page
(1,000 if \code{legacy} if \code{FALSE}).}

\item{endpoint}{Name of the API endpoint to search in, as listed on
\href{https://patentsview.org/apis/api-endpoints}{PatentsView}.}

\item{post}{Logical; if \code{FALSE}, will make a GET rather than POST request.}

\item{retry}{Maximum number of times to retry a request, in the event of a rate limit.}

\item{verbose}{Logical; if \code{FALSE}, does not print status messages.}

\item{legacy}{Logical; if \code{FALSE}, will use the beta API, which requires a key.}

\item{key}{Beta API key (\href{https://patentsview.org/apis/keyrequest}{Request a key});
defaults to the \code{PATENTSVIEW_KEY} environment variable.}

\item{cores}{Number of CPU cores to use when retrieving multiple pages of results.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites an existing result.}
}
\value{
A \code{data.frame} of results, if any were found. This may contain list entries
of \code{data.frame} with varying dimensions, depending on the \code{endpoint} and \code{fields}.
}
\description{
Submits a query to one of the \href{https://patentsview.org}{PatentsView} endpoints.
}
\examples{
\dontrun{
# search for patents with a meteorology classification:
# https://www.uspto.gov/web/patents/classification/cpc/html/cpc-G01W.html
results <- uspto_search_patentsview(list(cpc_group_id = "G01W"))

# search by date and keyword:
results <- uspto_search_patentsview(list(
  "_and" = list(
    list(patent_date = "2002-01-08"),
    list("_text_any" = list(patent_abstract = "motorcycle"))
  )
))
}
}
